Remove deprecated warnings
