+++ .changeset/famous-apes-exist.md
@@ -1,0 +1,6 @@
+---
+'@powersync/vue': minor
+'@powersync/react': major
+---
+
+Breaking: Remove `usePowerSyncQuery` and `usePowerSyncWatchedQuery` (use `useQuery` instead) and `usePowerSyncStatus` (use `useStatus` instead).
