+++ .changeset/five-roses-relax.md
@@ -1,0 +1,9 @@
+---
+'@powersync/react-native': major
+'@powersync/common': major
+'@powersync/web': major
+'@powersync/capacitor': minor
+'@powersync/tauri-plugin': patch
+---
+
+Remove `SQLOnChangeOptions.rawTableNames` and `WatchOnChangeHandler.onError` - both were already ignored before.
