+++ packages/common/etc/common.api.md
@@ -940,8 +940,6 @@ export type SqliteValue = string | number | bigint | number[] | Uint8Array | nul
 
 // @public (undocumented)
 export interface SQLOnChangeOptions {
-    // @deprecated (undocumented)
-    rawTableNames?: boolean;
     // (undocumented)
     signal?: AbortSignal;
     // (undocumented)
@@ -1353,8 +1351,6 @@ export interface WatchOnChangeEvent {
 export interface WatchOnChangeHandler {
     // (undocumented)
     onChange: (event: WatchOnChangeEvent) => Promise<void> | void;
-    // (undocumented)
-    onError?: (error: Error) => void;
 }
 
 // @alpha
