+++ packages/vue/src/index.ts
@@ -1,7 +1,4 @@
 export * from './composables/powerSync.js';
-export { usePowerSyncQuery } from './composables/usePowerSyncQuery.js';
-export { usePowerSyncStatus } from './composables/usePowerSyncStatus.js';
-export { usePowerSyncWatchedQuery } from './composables/usePowerSyncWatchedQuery.js';
 export { useQuery } from './composables/useQuery.js';
 export { useStatus } from './composables/useStatus.js';
 export { useWatchedQuerySubscription } from './composables/useWatchedQuerySubscription.js';
